feat(ensindexer): simplify ponder helpers
