+++ apps/ensindexer/src/indexing-status/index.ts
@@ -4,4 +4,4 @@ export {
   fetchPonderMetrics,
   fetchPonderStatus,
   DEFAULT_METRICS_FETCH_TIMEOUT,
-} from "./ponder-helpers";
+} from "./ponder-metadata";
