+++ packages/ensnode-sdk/src/shared/index.ts
@@ -3,6 +3,7 @@ export * from "./collections";
 export * from "./types";
 export * from "./serialized-types";
 export {
+  deserializeBlockNumber,
   deserializeBlockRef,
   deserializeChainId,
   deserializeDatetime,
